--- app/src/main/java/hiof/g12/compose/sceen/HobbiesScreen.kt
@@ -1,4 +1,0 @@
-package hiof.g12.compose.sceen
-
-class HobbiesScreen {
-}
