--- app/src/main/java/hiof/g12/compose/sceen/User/UserCalendarScreen.kt
@@ -1,4 +1,0 @@
-package hiof.g12.compose.sceen.User
-
-class UserCalendarScreen {
-}
