--- app/src/main/java/hiof/g12/compose/sceen/User/UserScreen.kt
@@ -1,4 +1,0 @@
-package hiof.g12.compose.sceen.User
-
-class UserScreen {
-}
