+++ app/src/main/java/hiof/g12/compose/screen/AddActivityScreen.java
@@ -1,4 +1,4 @@
-package hiof.g12.compose.sceen;
+package hiof.g12.compose.screen;
 
 public class AddActivityScreen {
 }
