+++ app/src/main/java/hiof/g12/compose/screen/HomeScreen.kt
@@ -1,0 +1,4 @@
+package hiof.g12.compose.screen
+
+class HomeScreen {
+}
