+++ app/src/main/java/hiof/g12/compose/screen/LoginScreen.kt
@@ -1,10 +1,9 @@
-package hiof.g12.compose.sceen
+package hiof.g12.compose.screen
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.heightIn
 import androidx.compose.material3.Surface
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
