+++ app/src/main/java/hiof/g12/compose/screen/User/UserCalendarScreen.kt
@@ -1,0 +1,4 @@
+package hiof.g12.compose.screen.User
+
+class UserCalendarScreen {
+}
