+++ app/src/main/java/hiof/g12/compose/screen/User/UserDiaryScreen.kt
@@ -1,0 +1,4 @@
+package hiof.g12.compose.screen.User
+
+class UserDiaryScreen {
+}
