+++ app/src/main/java/hiof/g12/compose/screen/WelcomeLayout.kt
@@ -1,7 +1,6 @@
-package hiof.g12.compose.sceen
+package hiof.g12.compose.screen
 
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
